CategoryItem  a separate component
