+++ Joker/src/components/ui/categories/index.tsx
@@ -1,0 +1,2 @@
+export * from './categoryItem';
+export * from './categories';
