+++ Joker/src/components/ui/index.ts
@@ -21,5 +21,6 @@ export * from './Tag';
 export * from './Avatar';
 export * from './TicketCardColumn';
 export * from './TicketCardRow';
-export * from './categories';
+export * from './categories/categories';
 export * from './headerList';
+export * from './separator';
